libdwP.h: Add locking to __libdw_dieabbrev
        * libdw/libdwP.h (__libdw_dieabbrev): Add locking.
        (struct Dwarf): Add dwarf_lock.
        (struct Dwarf_CU): Add abbrev_lock, split_lock.

Signed-off-by: Heather S. McIntyre <hsm2@rice.edu>
Signed-off-by: Aaron Merey <amerey@redhat.com>
Signed-off-by: Mark Wielaard <mark@klomp.org>

v2 changes:
This replaces patch "libdw: Add locking around __libdw_dieabbrev for
dwarf_hasattr".  Mark suggested that we remove lazy abbrev reading
in order to simplify the locking of __libdw_dieabbrev.  This is
a fair bit of work so for now lets just put a write lock around all
of __libdw_dieabbrev.  The removal of lazy reading of abbrev will be
done in a future patch.