eu-stacktrace WIP: optional LIBDWFL_TRACKS_UNWOUND_SOURCE patch
patch -p1 <libdwfl-report-unwound-source.patch

This is an optional diagnostic patch that tracks the unwinding
mechanism used to produce each Dwfl_Frame.

Plus some #ifdef code in src/stacktrace.c that uses the
field to report per-frame or per-process which unwinding
mechanism was used.

Pending to figure out a better way to expand the libdwfl
data structures with additional tracking info.