eu-stacktrace WIP: fix register order, 32-bit unwinding
The exciting 'scrambled eggs' transformations performed on the
x86 register files as it passes from hw to perf to DWARF are
quite something and require some magic numbers (sources carefully cited)
to be made to work correctly. The previous code worked deceptively
well but missed various cases.

h/t to qt perfparser for showing the precedent that this can be done,
even if our magic numbers are slightly different (due to omitting
segment registers).