eu-stacktrace [6/8]: use dwfl_process_tracker_find_pid
Initial minimal change to ensure dwfl_process_tracker_find_pid is
tested. For now, we keep the additional dwfltab code in stacktrace.c,
since it's used to track statistics.

Later, we will need to switch to storing statistics in
dwfl->process->callbacks_arg. This requires some additional design to
keep the statistics from being lost when a pid is reused and the
corresponding processtracker table entry is replaced.

* src/stacktrace.c (sysprof_init_dwfl): New function.
  (sysprof_find_dwfl): Rename the existing sysprof_init_dwfl.
  Also use dwfl_process_tracker_find_pid with callback.
  (sysprof_unwind_cb): Rename the existing sysprof_init_dwfl.