src/readelf.c: Support concurrency for -w, --debug-dump
Implement concurrent execution of print_debug_* functions during handling
of -w, --debug-dump using libthread.a.

A new `-C, --concurrency=NUM` command line option controls the maximum
number of threads that may be used. This value defaults to the number of CPUs.

Job output is buffered and printed in the order that jobs were added to
the queue. This helps preserve the existing order of stdout output. Full
support for output buffering in print_debug_* functions is added in the
next patch in this series.

	* src/readelf.c (default_concurrency): Function estimating the
	maximum number of threads.
	(parse_opt): Handle -C, --concurrency=NUM.
	(do_job): Entry point function for threads.
	(schedule_job): If thread safety is enabled, add job to the
	job queue.  Otherwise just run the job from the main thread.
	(print_debug): Pass print_debug_* function pointers and
	args to schedule_job. Also call run_jobs if thread safety
	is enabled.

Signed-off-by: Aaron Merey <amerey@redhat.com>

v3:
Replace assert before run_jobs with a conditional.