libdwfl [4/12]: add dwfl_perf_sample_preferred_regs_mask
Changes for v2:

- guard the linux/perf_events.h include properly with an #if defined __linux__

* * *

Since libebl is a private interface, subsequent patches in the series
introduce another api wrapping the libebl perf register handling.  In
this patch, add an interface to access the preferred set of registers
that libdwfl would like to see to allow proper unwinding of stack
sample data.

It feels like an implementation detail, but is unfortunately needed
because the profiler invoking perf_event_open and the elfutils
unwinding code need to agree on the set of registers to be
requested. So the profiler has to ask elfutils for this detail.

* libdwfl/libdwfl.h (dwfl_perf_sample_preferred_regs_mask):
  New function.
* libdwfl/dwfl_perf_frame.c: New file.
  (dwfl_perf_sample_preferred_regs_mask): New function.
* libdw/libdw.map: Add dwfl_perf_sample_preferred_regs_mask.
* libdwfl/Makefile.am: Add dwfl_perf_frame.c.