libdw: Make __libdw_fde_by_offset static in libdw/fde.c
__libdw_fde_by_offset is marked as an extern internal function in
libdw/cfi.h (a not public header file). But it is (now) only used in
libdw/fde.c, called from __libdw_find_fde. It was originally used for
dwarf_cfi_validate_fde, but that function was deleted and never made
public. So simplify things and make it static in libdw/fde.c.

	* libdw/cfi.h (__libdw_fde_by_offset): Remove.
	* libdw/fde.c (__libdw_fde_by_offset): Make static.