debuginfod-find.c: Avoid leaving temp files in the cache
Add a SIGINT handler to debuginfod-find so that ctrl-c during
a download causes clean up to occur instead of immediately killing
the process and possibly leaving temp files in the client cache.

This change requires setting the client progressfn unconditionally
instead of only when --verbose is given.  This in turn requires a small
change to a debuginfod-find testcase in run-debuginfod-artifact-running.sh
to account for output differences between the debuginfod-find progressfn
and the debuginfod_client default_progressfn which was previously used
when debuginfod-find lacked -v but had DEBUGINFOD_PROGRESS=1 set.

https://sourceware.org/bugzilla/show_bug.cgi?id=33163

Signed-off-by: Aaron Merey <amerey@redhat.com>