libelf: elf[32|64]_offscn shouldn't return a result for empty sections
offscn sets the result before checking the section isn't empty. It
assumes the result will be reset for the next section that matches the
given offset. But this might not be the case, for example if this was
the last section. It will then return that section (and set elf_errno)
instead of returning NULL to indicate no non-empty section matched.

	* libelf/elf32_offscn.c (offscn): Move assignment to result
	after empty size check.

Signed-off-by: Mark Wielaard <mark@klomp.org>